Introduce a ThreadManager
Create a simple thread manager class that uses ExecutorService and creates a map of "futures."
Allow it to take submitted jobs, manage the jobs, and return their status and provide an await termination timer. I have an isJobComplete method in case anything needs to "wait" in the future.
Use this in the BackupManager to execute tasks on their own threads. Wait 30 seconds for completion.